refactor (appwide): switch project structure to feature-first structure
* also do minor cleanups
* commonize common components

The feature-first approach demands that we create a new folder for every new feature that we add to our app. And inside that folder, we can add the layers themselves as sub-folders.
I find this approach more logical because we can easily see all the files that belong to a certain feature, grouped by layer.